refactor(test): move routing/test-helpers.js to TypeScript
