+++ packages/astro/test/units/routing/getstaticpaths-cache.test.ts
@@ -4,7 +4,7 @@ import type { ComponentInstance } from '../../../dist/types/astro.js';
 import type { AstroLogMessage, AstroLoggerDestination } from '../../../dist/core/logger/core.js';
 import { AstroLogger } from '../../../dist/core/logger/core.js';
 import { RouteCache, callGetStaticPaths } from '../../../dist/core/render/route-cache.js';
-import { dynamicPart, makeRoute } from './test-helpers.js';
+import { dynamicPart, makeRoute } from './test-helpers.ts';
 
 function mod(overrides: Partial<ComponentInstance>): ComponentInstance {
 	return overrides as ComponentInstance;
