+++ packages/astro/test/units/routing/preview-routing.test.ts
@@ -2,7 +2,7 @@ import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
 import type { AstroConfig } from '../../../dist/types/public/config.js';
 import { Router } from '../../../dist/core/routing/router.js';
-import { dynamicPart, makeRoute, staticPart } from './test-helpers.js';
+import { dynamicPart, makeRoute, staticPart } from './test-helpers.ts';
 
 describe('preview routing (unit)', () => {
 	const routes = (trailingSlash: AstroConfig['trailingSlash']) => [
