+++ packages/astro/test/units/routing/router-match.test.ts
@@ -1,7 +1,7 @@
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
 import { Router } from '../../../dist/core/routing/router.js';
-import { dynamicPart, makeRoute, spreadPart, staticPart } from './test-helpers.js';
+import { dynamicPart, makeRoute, spreadPart, staticPart } from './test-helpers.ts';
 
 describe('Router.match', () => {
 	it('prefers static routes over dynamic routes', () => {
