+++ packages/astro/test/units/routing/routing-priority.test.ts
@@ -1,7 +1,7 @@
 import assert from 'node:assert/strict';
 import { describe, it } from 'node:test';
 import { Router } from '../../../dist/core/routing/router.js';
-import { dynamicPart, makeRoute, spreadPart, staticPart } from './test-helpers.js';
+import { dynamicPart, makeRoute, spreadPart, staticPart } from './test-helpers.ts';
 
 const trailingSlash = 'ignore';
